requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
billiard==3.6.4.0
bleach==5.0.1
boto3==1.24.81
botocore==1.27.81
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-braces==1.15.0
django-crispy-forms==1.14.0
django-filter==22.1
django-model-utils==4.2.0
django-modeltranslation==0.18.4
django-request-logging==0.7.5
django-rest-framework==0.1.0
django-rest-registration==0.7.3
django-rosetta==0.9.8
django-star-ratings==0.9.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
kombu==5.2.4
loguru==0.6.0
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
polib==1.1.1
prompt-toolkit==3.0.31
pycodestyle==2.9.1
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-stdnum==1.17
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
swapper==1.3.0
toml==0.10.2
urllib3==1.26.12
validators==0.20.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
